Update template_editor.html
This statement {if condition="isset($value.max_chars)"}{$value.max_chars}{else}90{/if} makes the columns so large that it is impossible to view the ARI configurable column. The previous versions contain the value of '40'. I am adding this value back as it does not seem to break anything with our provisioner package for Yealink V70/V72, Cisco, or Polycom phones.