Update phonesettings.module
The statement method='GET' works when you do not have a lot of data to be updated. When we are using our remote phonebook setting with our Yealink V70 package with 120+ fields entered and customizable, we the GET method fails due to the URL being displayed too long. When we use the POST method, there are no errors.